requirements.txt
asgiref==3.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.0
pymongo==3.12.3
python-snappy==0.6.0
pytz==2021.3
sqlparse==0.2.4
tzdata==2021.5
winkerberos==0.8.0
